move analysis/transforms out of `fpy/ast`
