+++ fpy2/analysis/__init__.py
@@ -1,3 +1,5 @@
 """Program analyses for FPy programs"""
 
 from .define_use import DefineUse, DefineUseAnalysis, Definition, DefinitionCtx
+from .live_vars import LiveVars
+from .syntax_check import SyntaxCheck
